requirements.txt
# Core Streamlit framework
streamlit>=1.28.0

# Google Generative AI (Gemini)
google-generativeai>=0.3.0

# LangChain components
langchain>=0.1.0
langchain-community>=0.0.20

# Vector database and embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0

# PDF processing
pypdf>=3.17.0

# Additional utilities
python-dotenv>=1.0.0
typing-extensions>=4.8.0

# For better performance
numpy>=1.24.0
tiktoken>=0.5.0